C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik4.cpp -o C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik4.exe  -Wshadow -Winit-self -Wredundant-decls -Wcast-align -Wundef -Wfloat-equal -Winline -Wunreachable-code -Wmissing-declarations -Wmissing-include-dirs -Wswitch-enum -Wswitch-default -Weffc++ -Wmain -Wextra -Wall -g -pipe -fexceptions -Wcast-qual -Wconversion -Wctor-dtor-privacy -Wempty-body -Wformat-security -Wformat=2 -Wignored-qualifiers -Wlogical-op -Wmissing-field-initializers -Wnon-virtual-dtor -Woverloaded-virtual -Wpointer-arith -Wsign-promo -Wstack-usage=8192 -Wstrict-aliasing -Wstrict-null-sentinel -Wtype-limits -Wwrite-strings -D_DEBUG -D_EJUDGE_CLIENT_SIDE -D_EJC

+++ Ozhaev Zaynudin Sharik4.cpp
@@ -16,10 +16,10 @@ int main()
 
 void MoveBall()
     {
-    float x1 = W/4;
-    float y1 = H/2;
-    float x2 = W/4*3;
-    float y2 = H/2;
+    int x1 = W/4;
+    int y1 = H/2;
+    int x2 = W/4*3;
+    int y2 = H/2;
     int r1 = 20;
     int r2 = 50;
     int vx1 = rand() % 10;
